Supress "supported by" if unsupported
Some content (like heat templates or glance
images) will be added to the catalog without
anyone supporting them. In this case, we
will leave "supported by" out of the entry in
the YAML - then if there is no supported by
in the YAML, we should not show "Supported
by" when rendering the page.

Change-Id: I5543cfc973916d80b75c767808830c48ac34d1c5