Fix Debian link on landing page
The link on the landing page pointed to a
Debian glance image. Since that entry was
added to the landing page, the associated
Debian entry in the glance section was updated
to be more specific (which broke the link
on the landing page.

This commit fixes that link on the landing
page.

Change-Id: Ief25f2d9ad80c4fe4fb880aa4e2ceff3daf02abd