Update GA tracking code
The OpenStack foundation maintains a Google Analytics
account for tracking access. The stats from this
account will be shared regularly with docaedo,
who will then share these stats with the appropriate
MLs (similar to how this has been handled by the
OpenStack docs team).

This commit updates the tracking code so visits to
the App Catalog are properly tracked by using a new
sub-code for GA and also sending anchor tags along
with the clicks.

Closes-Bug: 1461167

Change-Id: Ie87f6da290718e8b8dcc72f65f588161f923c03d