Combine multiple YAMLs into single YAML
This commit combines the contents of the
three YAML files (for Glance, Heat and
Murano assets) into a single file
called assets.yaml. It also modifies
the javascript for the website so the
site will load the assets.yaml file
instead of the three individual YAML
files.

Change-Id: Ia004e6d4f3b51ca853945b136abdc1599d7af4b8
Implements: blueprint expand-asset-types