WIP - Python Server
This switches the server over to being Python rather then just
doing static files. This allows for:
 * Dynamic API.
 * Faster loads due to supporting compression of Javascript.
 * Javascript/CSS from local webserver rather then remote.
 * XStatic managed Javascript/CSS.
 * Sharing code between the app-catalog and Horizon.

This patch set enables most of the infrastructure to do that
and copies in the initial website. Minimal changes have been
made to the site to make it compatible with the new code.

Further reviews will switch the site to use more of the new
functionality.

Change-Id: I2fa9686d3e5db4ee43a02aa7d09b998472e02bf6