WIP - Pythonserver cleanup #1
Add SCSS and pull in a lot of dependencies directly into the server
via XStatic rather then link externally. Update testicons.html to
use it to create a testing ground to sharing code between the python
server and the horizon plugin.

This WIP pulls in parts of Horizon. Ideally those would be put in a
common library.

Change-Id: I80324c9c58c255d6afd1c489172c703d9c9666f1