Skip assets with active==false attribute
This commit modifies the yaml2json script to cause it to skip entries
in the assets YAML if the asset includes an attribute called "active"
set to false.

At this time there are no entries with anything in the
attributes/active field.  In the future a script will run daily to
check for stale entries - any found will result in a review being
submitted which adds "active: false" to the attributes section,
which will cause that asset to no longer be displayed. The entry will
remain however so it can quickly be adjusted by setting active to true
(or removing that field) if the link to the asset is repaired.

Change-Id: I3698cd0c038da0fc130b905303c3df4206c8b034
Implements: blueprint detect-stale-entries