Move and keep assets_merge.yaml
This commit changes where assets_merge.yaml file is created, and does
not delete that file after update_assets.sh is run. The file will now
be created in the /tmp directory where it will not cause git to think
any new files have been added in subsequent pull events, and also
leaves this file in place so someone from the infra team can SSH to
the box and validate whether or not the fields are being populated
correctly on the production server.

This will not resolve the linked bug, but will be a big aid in
tracking down the cause.

Change-Id: I9c9f3f80b5e702e6c259787df741a5f963286688
Partial-Bug: 1553572