Keep assets_merge.yaml
This commit removes the line that deleted assets_merge.yaml.  This
file is left in place so someone from the infra team can SSH to the
box and validate whether or not the fields are being populated
correctly on the production server.

This will not resolve the linked bug, but will be a big aid in
tracking down the cause.

Change-Id: I9c9f3f80b5e702e6c259787df741a5f963286688
Partial-Bug: 1553572