Use direct links for logos
Previously in patch https://review.openstack.org/#/c/362916 we used
links which point on logog file in repo.
Now it looks like it's not correct format, because logos still are not
displayed in App Catalog.
This patch set urls to directly to images.
Also fix wrongly used Nexus log for ELK.

Change-Id: I8f61b91b4bbe127e7f1ab9b06375d989f3ee4ceb